Add 'gcoap.rs', refactor
